% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{acme_options}
\alias{acme_options}
\title{Global ACME options}
\usage{
acme_options(verbose = NULL, server = NULL)
}
\arguments{
\item{verbose}{emit verbose http output}

\item{server}{acme server endpoint root}
}
\description{
Get and set global ACME options.
}
