\name{example.agff}
\alias{example.agff}
\docType{data}
\title{An example ACME data structure of class aGFF}
\description{
  An aGFF data structure from two Nimblegen arrays, custom tiled to
  include multiple HOX genes.
}
\usage{data(example.agff)}
\format{
  The format is:
 chr "example.agff"
}
\source{
  From Scacheri et al., Plot Genet, 2006.  Pubmed ID 16604156
}
\examples{
data(example.agff)
example.agff
}
\keyword{datasets}
